+++ src/hive_mariachi/main/engine.h
@@ -265,7 +265,7 @@ namespace mariachi {
             void removeMainThreadStage(stages::Stage *stage);
             void addTask(tasks::Task *task);
             void removeTask(tasks::Task *task);
-            void startCameraTransition(const std::string &cameraName, float duration);
+            void startCameraTransition(const std::string &cameraName, int duration);
             void cancelCameraTransition();
             void getCurrentProcessIdString(std::string &currentProcessIdString);
             stages::StageRunner *getStageRunner(stages::Stage *stage);
